+++ integration/local.py
@@ -117,6 +117,8 @@ def run_lila():
         [
             "docker",
             "run",
+            "--pull",
+            "always",
             "--name",
             BDIT_LILA,
             "--network",