+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ To be released
 * Added ``client.relations.block`` and ``client.relations.unblock`` for blocking/unblocking users.
 * Implemented ``/api/games/export/imports`` under
   ``client.games.export_imported``.
+* Implemented ``/api/study/by/{username}`` under
+  ``client.studies.get_by_user`` to get the metadata of all studies
+  of a given user.
 
 Thanks to all the contributors who helped to this release:
 - @hsheth2